Validating fields

+++ src/routes.js
@@ -1,8 +1,19 @@
 import { Database } from './database.js';
 import { randomUUID } from 'node:crypto'
 import { buildRoutePath } from './utils/build-route-path.js';
+import { httpValidator } from './middlewares/http-validator.js';
 
 const database = new Database()
+
+function validateFields(res, title, description) {
+    if (!title) {
+        return httpValidator(res, 400, "Título inválido")
+    }
+
+    if (!description) {
+        return httpValidator(res, 400, "Descrição inválida")
+    }
+}
 
 export const routes = [
     {
@@ -25,6 +36,8 @@ export const routes = [
         handler: (req, res) => {
             const { title, description } = req.body
 
+            validateFields(res, title, description)
+
             const task = {
                 id: randomUUID(),
                 title,
@@ -46,6 +59,13 @@ export const routes = [
         handler: (req, res) => {
             const { id } = req.params
             const { title, description } = req.body
+
+            if (!database.isIdValid('tasks', id)) {
+                return httpValidator(res, 400, "ID inexistênte")
+            }
+
+            validateFields(res, title, description)
+
             database.update('tasks', id, { title, description })
             return res.writeHead(204).end()
 
@@ -56,6 +76,11 @@ export const routes = [
         path: buildRoutePath('/tasks/:id/complete'),
         handler: (req, res) => {
             const { id } = req.params
+
+            if (!database.isIdValid('tasks', id)) {
+                return httpValidator(res, 400, "ID inexistênte")
+            }
+
             database.setCompleted('tasks', id)
             return res.writeHead(204).end()
         }
@@ -65,6 +90,11 @@ export const routes = [
         path: buildRoutePath('/tasks/:id'),
         handler: (req, res) => {
             const { id } = req.params
+
+            if (!database.isIdValid('tasks', id)) {
+                return httpValidator(res, 400, "ID inexistênte")
+            }
+
             database.delete('tasks', id)
             return res.writeHead(204).end()
         }
